nanny/requirements.txt
hazelcast-python-client
flask
flask-classy
requests
python-dotenv
docker
pymongo